modernized repo

+++ src/pysmile/image_conv_util.py
@@ -1,8 +1,8 @@
-
 from PIL import Image
 import numpy as np
 
 # http://stackoverflow.com/questions/9166400/convert-rgba-png-to-rgb-with-pil
+
 
 def alpha_to_color(image, color=(255, 255, 255)):
     """Set all fully transparent pixels of an RGBA image to the specified color.
@@ -25,7 +25,7 @@ def alpha_to_color(image, color=(255, 255, 255)):
     g[a == 0] = color[1]
     b[a == 0] = color[2]
     x = np.dstack([r, g, b, a])
-    return Image.fromarray(x, 'RGBA')
+    return Image.fromarray(x, "RGBA")
 
 
 def alpha_composite(front, back):
@@ -40,20 +40,22 @@ def alpha_composite(front, back):
     """
     front = np.asarray(front)
     back = np.asarray(back)
-    result = np.empty(front.shape, dtype='float')
+    result = np.empty(front.shape, dtype="float")
     alpha = np.index_exp[:, :, 3:]
     rgb = np.index_exp[:, :, :3]
     falpha = front[alpha] / 255.0
     balpha = back[alpha] / 255.0
     result[alpha] = falpha + balpha * (1 - falpha)
-    old_setting = np.seterr(invalid='ignore')
-    result[rgb] = (front[rgb] * falpha + back[rgb] * balpha * (1 - falpha)) / result[alpha]
+    old_setting = np.seterr(invalid="ignore")
+    result[rgb] = (front[rgb] * falpha + back[rgb] * balpha * (1 - falpha)) / result[
+        alpha
+    ]
     np.seterr(**old_setting)
     result[alpha] *= 255
     np.clip(result, 0, 255)
     # astype('uint8') maps np.nan and np.inf to 0
-    result = result.astype('uint8')
-    result = Image.fromarray(result, 'RGBA')
+    result = result.astype("uint8")
+    result = Image.fromarray(result, "RGBA")
     return result
 
 
@@ -66,7 +68,7 @@ def alpha_composite_with_color(image, color=(255, 255, 255)):
     color -- Tuple r, g, b (default 255, 255, 255)
 
     """
-    back = Image.new('RGBA', size=image.size, color=color + (255,))
+    back = Image.new("RGBA", size=image.size, color=color + (255,))
     return alpha_composite(image, back)
 
 
@@ -84,6 +86,7 @@ def pure_pil_alpha_to_color_v1(image, color=(255, 255, 255)):
     color -- Tuple r, g, b (default 255, 255, 255)
 
     """
+
     def blend_value(back, front, a):
         return (front * a + back * (255 - a)) / 255
 
@@ -99,6 +102,7 @@ def pure_pil_alpha_to_color_v1(image, color=(255, 255, 255)):
 
     return im
 
+
 def pure_pil_alpha_to_color_v2(image, color=(255, 255, 255)):
     """Alpha composite an RGBA Image with a specified color.
 
@@ -112,9 +116,10 @@ def pure_pil_alpha_to_color_v2(image, color=(255, 255, 255)):
 
     """
     image.load()  # needed for split()
-    background = Image.new('RGB', image.size, color)
+    background = Image.new("RGB", image.size, color)
     background.paste(image, mask=image.split()[3])  # 3 is the alpha channel
     return background
+
 
 def convert_to_palette(image):
     """Convert transparent image to GIF
@@ -128,7 +133,7 @@ def convert_to_palette(image):
     # To have best quality, have non-destructive RGBA to RGB conversion first
     background = pure_pil_alpha_to_color_v2(image)
     # Convert to palette format
-    background = background.convert('P', palette=Image.ADAPTIVE, colors=255)
+    background = background.convert("P", palette=Image.ADAPTIVE, colors=255)
     # Set all pixel values below 128 to 255, and the rest to 0
     mask = Image.eval(alpha, lambda a: 255 if a <= 128 else 0)
     # Paste the color of index 255 and use alpha as a mask
